Applied recommendatinos
